Struct wasmer_argus::argus::Argus

pub struct Argus {
    pub config: ArgusConfig,
    pub client: WasmerClient,
}

Fields
config: ArgusConfigclient: WasmerClient
Implementations

impl Argus

pub async fn fetch_packages( &self, s: UnboundedSender<PackageVersionWithPackage>, p: ProgressBar, config: Arc<ArgusConfig> ) -> Result<()>

Fetch all packages from the registry

pub(crate) async fn download_package<'a>( test_id: u64, path: &'a PathBuf, url: &'a Url, p: &'a ProgressBar ) -> Result<()>

pub async fn get_path( config: Arc<ArgusConfig>, pkg: &PackageVersionWithPackage ) -> PathBuf

Return the complete path to the folder of the test for the package, from the outdir to the hash

pub fn get_package_id(pkg: &PackageVersionWithPackage) -> String

impl Argus

pub async fn run(self) -> Result<()>

Start the testsuite using the configuration in Self::config

async fn test( test_id: u64, config: Arc<ArgusConfig>, package: &PackageVersionWithPackage, p: ProgressBar ) -> Result<()>

Perform the test for a single package

async fn to_test(&self, pkg: &PackageVersionWithPackage) -> bool

Checks whether or not the package should be tested

async fn write_report(path: &Path, result: TestReport) -> Result<()>